requirements.txt
python-dotenv
google-genai
transformers
fastapi
uvicorn
requests
pymupdf

